requirements.txt
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-ai-inference==1.0.0b9
azure-common==1.1.28
azure-core==1.32.0
azure-functions==1.24.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
azurefunctions-extensions-base==1.1.0
azurefunctions-extensions-bindings-blob==1.1.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fitz==0.0.1.dev2
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jupyter_client==8.6.3
jupyter_core==5.9.1
Levenshtein==0.27.3
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.59.7
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
prov==2.1.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.5
pyodbc==5.3.0
pyparsing==3.2.5
pypdfium2==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
pyzmq==27.1.0
RapidFuzz==3.14.3
rdflib==7.4.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
xlsxwriter==3.2.9
yarl==1.22.0
